[popover][tooltip] Inline hooks
