+++ packages/react/src/popover/positioner/PopoverPositionerContext.ts
@@ -1,18 +1,27 @@
 'use client';
 import * as React from 'react';
-import { usePopoverPositioner } from './usePopoverPositioner';
+import type { Side, Align } from '../../utils/useAnchorPositioning';
+import type { FloatingContext } from '../../floating-ui-react';
 
-export const PopoverPositionerContext = React.createContext<
-  usePopoverPositioner.ReturnValue | undefined
->(undefined);
+export interface PopoverPositionerContext {
+  side: Side;
+  align: Align;
+  arrowRef: React.RefObject<Element | null>;
+  arrowUncentered: boolean;
+  arrowStyles: React.CSSProperties;
+  context: FloatingContext;
+}
+
+export const PopoverPositionerContext = React.createContext<PopoverPositionerContext | undefined>(
+  undefined,
+);
 
 export function usePopoverPositionerContext() {
   const context = React.useContext(PopoverPositionerContext);
-  if (context === undefined) {
+  if (!context) {
     throw new Error(
       'Base UI: PopoverPositionerContext is missing. PopoverPositioner parts must be placed within <Popover.Positioner>.',
     );
   }
-
   return context;
 }
